refactor + remove CATE fitting functionality
